ref: convert sentry settings import hook to a normal module (#60811)
this speeds up test discovery by about 1% (discovery is import-heavy and
every import went through the custom meta_path loader)

closes #60799

<!-- Describe your PR here. -->
