+++ src/sentry/runner/initializer.py
@@ -19,6 +19,15 @@ logger = logging.getLogger("sentry.runner.initializer")
 T = TypeVar("T")
 
 
+class ConfigurationError(ValueError, click.ClickException):
+    def show(self, file=None):
+        if file is None:
+            from click._compat import get_text_stderr
+
+            file = get_text_stderr()
+        click.secho(f"!! Configuration error: {self!r}", file=file, fg="red")
+
+
 def register_plugins(settings: Any, raise_on_plugin_load_failure: bool = False) -> None:
     from sentry.plugins.base import plugins
 
@@ -176,8 +185,6 @@ def bootstrap_options(settings: Any, config: str | None = None) -> None:
             # Gracefully fail if yaml file doesn't exist
             options = {}
         except (AttributeError, ParserError, ScannerError) as e:
-            from .importer import ConfigurationError
-
             raise ConfigurationError("Malformed config.yml file: %s" % str(e))
 
         # Empty options file, so fail gracefully
@@ -185,8 +192,6 @@ def bootstrap_options(settings: Any, config: str | None = None) -> None:
             options = {}
         # Options needs to be a dict
         elif not isinstance(options, dict):
-            from .importer import ConfigurationError
-
             raise ConfigurationError("Malformed config.yml file")
     else:
         options = {}
@@ -418,8 +423,6 @@ def setup_services(validate: bool = True) -> None:
         tsdb,
     )
 
-    from .importer import ConfigurationError
-
     service_list = (
         analytics,
         buffer,
@@ -629,8 +632,6 @@ def apply_legacy_settings(settings: Any) -> None:
     # this is the only value that should prevent the app from booting up. Currently FLAG_REQUIRED is used to
     # trigger the Installation Wizard, not abort startup.
     if not settings.SENTRY_OPTIONS.get("system.secret-key"):
-        from .importer import ConfigurationError
-
         raise ConfigurationError(
             "`system.secret-key` MUST be set. Use 'sentry config generate-secret-key' to get one."
         )
@@ -684,8 +685,6 @@ def validate_snuba() -> None:
             snuba_enabled_features.add(feature)
 
     if snuba_enabled_features and not eventstream_is_snuba:
-        from .importer import ConfigurationError
-
         show_big_error(
             """
 You have features enabled which require Snuba,
@@ -701,8 +700,6 @@ See: https://github.com/getsentry/snuba#sentry--snuba
         raise ConfigurationError("Cannot continue without Snuba configured.")
 
     if not eventstream_is_snuba:
-        from .importer import ConfigurationError
-
         show_big_error(
             """
 It appears that you are requiring Snuba,
